move feature details to legacy file
